folder structure improved, redux toolkit thunk separation and renamed some folder name
