+++ src/App.jsx
@@ -1,8 +1,8 @@
 import { Routes, Route, Navigate } from "react-router-dom";
 import { useSelector } from "react-redux";
-import Dashboard from "./Components/Dashboard";
-import Login from "./Components/Login";
-import PrivateRoute from "./Components/PrivateRoute";
+import Dashboard from "./pages/Dashboard";
+import Login from "./pages/Login";
+import PrivateRoute from "./components/PrivateRoute";
 
 function App() {
   const { isLoggedIn } = useSelector((state) => state.auth);
