+++ src/reduxToolkit/auth/authSlice.js
@@ -1,0 +1,35 @@
+import { createSlice } from "@reduxjs/toolkit";
+import { loginAsync } from "./authThunk";
+
+const authSlice = createSlice({
+  name: "auth",
+  initialState: {
+    user: JSON.parse(localStorage.getItem("user")) || null,
+    isLoggedIn: localStorage.getItem("isLoggedIn") === "true",
+    error: null,
+  },
+  reducers: {
+    logout: (state) => {
+      state.user = null;
+      state.isLoggedIn = false;
+      state.error = null;
+
+      localStorage.removeItem("user");
+      localStorage.removeItem("isLoggedIn");
+    },
+  },
+  extraReducers: (builder) => {
+    builder
+      .addCase(loginAsync.fulfilled, (state, action) => {
+        state.user = action.payload;
+        state.isLoggedIn = true;
+        state.error = null;
+      })
+      .addCase(loginAsync.rejected, (state, action) => {
+        state.error = action.payload;
+      });
+  },
+});
+
+export const { logout } = authSlice.actions;
+export default authSlice.reducer;
